feat:  优化-资源管理下的业务选择器
--story=118133962
